zephyr: refactor native_sim tests into shared pytest modules
Move the shared Zephyr sample and OTA tests out of the
examples/zephyr/pytest directory into the plugin. Use per-example pytest
folder to include them.

Generalize ota_harness.py so the OTA fixture can be used for both
native_sim (generated image) and samples (compiled image).
