+++ .github/workflows/test-nsim.yml
@@ -99,7 +99,8 @@ jobs:
             -p nrf52_bsim/native                                                  \
             -T pouch/examples/zephyr/                                             \
             --pytest-args="--api-url=${{ inputs.api_url }}"                       \
-            --pytest-args="--api-key=${{ secrets[inputs.api_key_id] }}"
+            --pytest-args="--api-key=${{ secrets[inputs.api_key_id] }}"           \
+            --pytest-args="--ota-dummy-binary"
 
       - name: Safe upload twister artifacts
         if: success() || failure()
