+++ examples/zephyr/http_client/pytest/test_sample.py
@@ -1,0 +1,7 @@
+#
+# Copyright (c) 2026 Golioth, Inc.
+#
+# SPDX-License-Identifier: Apache-2.0
+#
+
+from pytest_pouch.zephyr_settings_tests import *  # noqa: F403
